chore: move DbBlockHeader to repository package
