+++ domains/headers.go
@@ -2,7 +2,6 @@ package domains
 
 import (
 	"math/big"
-	"strconv"
 	"time"
 
 	"github.com/libsv/bitcoin-hc/internal/chaincfg"
@@ -41,22 +40,6 @@ type BlockHeader struct {
 	CumulatedWork    *big.Int       `json:"work"`
 	PreviousBlock    chainhash.Hash `json:"prevBlockHash"`
 	DifficultyTarget uint32         `json:"difficultyTarget"`
-}
-
-// DbBlockHeader represent header saved in db.
-type DbBlockHeader struct {
-	Height           int32     `db:"height"`
-	Hash             string    `db:"hash"`
-	Version          int32     `db:"version"`
-	MerkleRoot       string    `db:"merkleroot"`
-	Timestamp        time.Time `db:"timestamp"`
-	Bits             uint32    `db:"bits"`
-	Nonce            uint32    `db:"nonce"`
-	State            string    `db:"header_state"`
-	Chainwork        string    `db:"chainwork"`
-	CumulatedWork    string    `db:"cumulatedWork"`
-	PreviousBlock    string    `db:"previousblock"`
-	DifficultyTarget uint32    `db:"difficultytarget"`
 }
 
 // HeaderArgs are sued to retrieve a single block header.
@@ -164,75 +147,6 @@ func (bh *BlockHeader) IsOrphan() bool {
 // IsLongestChain is the block part of the longest chain.
 func (bh *BlockHeader) IsLongestChain() bool {
 	return bh.State == LongestChain
-}
-
-// ConvertToBlockHeader converts one or whole slice of DbBlockHeaders to BlockHeaders
-// used after getting records from db.
-func ConvertToBlockHeader(dbBlockHeaders []*DbBlockHeader) []*BlockHeader {
-	if dbBlockHeaders != nil {
-		var blockHeaders []*BlockHeader
-
-		for _, header := range dbBlockHeaders {
-			h := header.ToBlockHeader()
-			blockHeaders = append(blockHeaders, h)
-		}
-		return blockHeaders
-	}
-	return nil
-}
-
-// ToBlockHeader converts work from string to big.Int and return BlockHeader.
-func (dbh *DbBlockHeader) ToBlockHeader() *BlockHeader {
-	if dbh.CumulatedWork == "" {
-		dbh.CumulatedWork = "0"
-	}
-	cumulatedWork, ok := new(big.Int).SetString(dbh.CumulatedWork, 10)
-	if !ok {
-		cumulatedWork = big.NewInt(0)
-	}
-
-	chainWork, err := strconv.ParseUint(dbh.Chainwork, 10, 64)
-	if err != nil {
-		chainWork = 0
-	}
-
-	hash, _ := chainhash.NewHashFromStr(dbh.Hash)
-	merkleTree, _ := chainhash.NewHashFromStr(dbh.MerkleRoot)
-	prevBlock, _ := chainhash.NewHashFromStr(dbh.PreviousBlock)
-
-	return &BlockHeader{
-		Height:           dbh.Height,
-		Hash:             *hash,
-		Version:          dbh.Version,
-		MerkleRoot:       *merkleTree,
-		Timestamp:        dbh.Timestamp,
-		Bits:             dbh.Bits,
-		Nonce:            dbh.Nonce,
-		Chainwork:        chainWork,
-		CumulatedWork:    cumulatedWork,
-		State:            HeaderState(dbh.State),
-		PreviousBlock:    *prevBlock,
-		DifficultyTarget: dbh.DifficultyTarget,
-	}
-}
-
-// ToDbBlockHeader converts BlockHeader to DbBlockHeader
-// used mainly to prepare record befor saving in db.
-func (bh BlockHeader) ToDbBlockHeader() DbBlockHeader {
-	return DbBlockHeader{
-		Height:           bh.Height,
-		Hash:             bh.Hash.String(),
-		Version:          bh.Version,
-		MerkleRoot:       bh.MerkleRoot.String(),
-		Timestamp:        bh.Timestamp,
-		Bits:             bh.Bits,
-		Nonce:            bh.Nonce,
-		State:            bh.State.String(),
-		Chainwork:        strconv.FormatUint(bh.Chainwork, 10),
-		CumulatedWork:    bh.CumulatedWork.String(),
-		PreviousBlock:    bh.PreviousBlock.String(),
-		DifficultyTarget: bh.DifficultyTarget,
-	}
 }
 
 // WrapWithHeaderState wraps BlockHeader with additional information creating BlockHeaderState.
